delete dead code
